Nextjs-Vite: Share URL-safe base64 helpers across font and image plugins
Both virtual-module IDs need the same encode/decode, so keep one copy.
